Fix Theme Script Server/Client Error
